修复一些部署方面的问题

+++ gen-config.mjs
@@ -1,4 +1,6 @@
 import fs from 'fs';
 
-const initialConfig = fs.readFileSync('config.yml.template').toString();
-fs.writeFileSync('config.yml', initialConfig);
+if (!fs.existsSync('config.yml')) {
+    const initialConfig = fs.readFileSync('config.yml.template').toString();
+    fs.writeFileSync('config.yml', initialConfig);
+}
